Post caching

+++ src/main/java/dev/neubert/backendsystems/socialmedia/adapters/in/api/controllers/PostWebController.java
@@ -7,14 +7,11 @@ import dev.neubert.backendsystems.socialmedia.application.domain.fakers.PostFake
 import dev.neubert.backendsystems.socialmedia.application.domain.mapper.PostMapper;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import jakarta.inject.Inject;
-import jakarta.validation.Valid;
+import jakarta.validation.Valid;test: checks empty body
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.ws.rs.*;
-import jakarta.ws.rs.core.HttpHeaders;
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
-import jakarta.ws.rs.core.UriInfo;
+import jakarta.ws.rs.core.*;
 import org.mapstruct.factory.Mappers;
 
 import java.util.stream.Collectors;
@@ -32,6 +29,7 @@ public class PostWebController {
 
     private UriInfo uriInfo;
     private HttpHeaders httpHeaders;
+    private CacheControl cacheControl;
 
 
     @GET
@@ -49,17 +47,19 @@ public class PostWebController {
             @QueryParam("size")
             long size
     ) {
-        var allPosts = this.postAdapter.readAllPosts();
 
+        setCacheControlFiveMinutes();
+        var allPosts = this.postAdapter.readAllPosts();
         var filteredPosts =
                 allPosts.stream().filter(post -> post.getContent().contains(query)).toList();
 
-        var paginatedPosts =
+        var result =
                 filteredPosts.stream().skip(offset).limit(size).collect(Collectors.toList());
 
         return Response.status(HttpResponseStatus.OK.code())
-                       .header("X-Total-Count", paginatedPosts.size())
-                       .entity(paginatedPosts)
+                       .header("X-Total-Count", result.size())
+                       .cacheControl(this.cacheControl).
+                       .entity(result)
                        .build();
     }
 
@@ -72,8 +72,14 @@ public class PostWebController {
             @PathParam("id")
             long id
     ) {
+        setCacheControlFiveMinutes();
         var requestedPost = this.postAdapter.getPostById(id);
-        return Response.ok(requestedPost).build();
+        if (requestedPost == null) {
+            return Response.status(HttpResponseStatus.NOT_FOUND.code()).build();
+        }
+        return Response.ok(requestedPost).tag(Long.toString(requestedPost.hashCode()))
+                       .cacheControl(this.cacheControl)
+                       .build();
     }
 
 
@@ -83,9 +89,12 @@ public class PostWebController {
             @Valid
             CreatePostDto model
     ) {
+        setCacheControlFiveMinutes();
         var result = this.postAdapter.createPost(model);
         return Response.status(HttpResponseStatus.CREATED.code())
                        .header("Location", createLocationHeader(result))
+                       .tag(Long.toString(result.hashCode()))
+                       .cacheControl(this.cacheControl)
                        .build();
     }
 
@@ -100,9 +109,15 @@ public class PostWebController {
             @Valid
             PostDto model
     ) {
+        setCacheControlFiveMinutes();
+        if (this.postAdapter.getPostById(id) == null) {
+            return Response.status(HttpResponseStatus.NOT_FOUND.code()).build();
+        }
         var result = this.postAdapter.updatePost(id, model);
         return Response.status(HttpResponseStatus.NO_CONTENT.code())
                        .header("Location", createLocationHeader(result))
+                       .tag(Long.toString(result.hashCode()))
+                       .cacheControl(this.cacheControl)
                        .build();
     }
 
@@ -115,6 +130,9 @@ public class PostWebController {
             long id
     ) {
         var toBeDeleted = this.postAdapter.getPostById(id);
+        if (toBeDeleted == null) {
+            return Response.status(HttpResponseStatus.NOT_FOUND.code()).build();
+        }
         if (this.postAdapter.deletePost(toBeDeleted)) {
             return Response.status(HttpResponseStatus.NO_CONTENT.code()).build();
         } else {
@@ -125,14 +143,24 @@ public class PostWebController {
     @POST
     @Path("populate")
     public Response populateDatabase() {
-        this.postAdapter.createPost(postMapper.postDtoToCreatePostDto(
+        setCacheControlFiveMinutes();
+        var result = this.postAdapter.createPost(postMapper.postDtoToCreatePostDto(
                 postMapper.postToPostDto(postFaker.createModel())));
-        return Response.status(HttpResponseStatus.CREATED.code()).build();
+        return Response.status(HttpResponseStatus.CREATED.code())
+                       .header("Location", createLocationHeader(result))
+                       .tag(Long.toString(result.hashCode()))
+                       .cacheControl(this.cacheControl)
+                       .build();
     }
 
 
     private String createLocationHeader(PostDto model) {
         return uriInfo.getRequestUriBuilder().path(Long.toString(model.getId())).build().toString();
+    }
+
+    private void setCacheControlFiveMinutes() {
+        this.cacheControl = new CacheControl();
+        this.cacheControl.setMaxAge(300);
     }
 
 }
